Make explicit long long

+++ src/cmd/ksh93/bltins/enum.c
@@ -200,10 +200,10 @@ static char* get_enum(Namval_t* np, Namfun_t *fp)
 {
 	static char buff[21];
 	struct Enum *ep = (struct Enum*)fp;
-	intmax_t n = nv_getn(np,fp);
+	long long n = nv_getn(np,fp);
 	if(n < ep->nelem)
 		return (char*)ep->values[n];
-	sfsprintf(buff,sizeof(buff),"%jd%c",n,0);
+	sfsprintf(buff,sizeof(buff),"%lld%c",n,0);
 	return buff;
 }
